requirements.txt
h5py==2.9.0
ortools==8.0.8283
pyyaml==5.1.2
numpy==1.17.3
torch==1.1.0
torch-scatter==1.4.0
torch-sparse==0.4.3
torch-cluster==1.4.5
torch-geometric==1.3.2
